extract reusable modal field component
