+++ packages/web/src/pages/upload-page/fields/ReleaseDateModalField.module.css
@@ -1,15 +1,3 @@
-/* Form display component */
-.displayTile {
-  display: flex;
-  flex-direction: column;
-  align-items: flex-start;
-  padding: var(--unit-4) var(--unit-6);
-  gap: var(--unit-4);
-
-  border: 1px solid var(--neutral-light-8);
-  border-radius: 8px;
-}
-
 .header {
   display: flex;
   width: 100%;
@@ -25,25 +13,27 @@
   line-height: 20px;
 }
 
-.caret {
-  color: var(--neutral-light-4);
-  height: 14px;
-  width: 14px;
-}
-
 .description {
   font-weight: var(--font-medium);
 }
 
+.preview {
+  display: flex;
+  flex-direction: column;
+  gap: var(--unit-2);
+}
+
 .valueDisplay {
+  box-sizing: content-box;
   display: flex;
   flex-direction: row;
   align-items: center;
   padding: 8px;
-  gap: 4px;
   background: var(--neutral-light-8);
   border: 1px solid var(--neutral-light-7);
   border-radius: 2px;
+  margin-top: var(--unit-2);
+  gap: var(--unit-2);
 
   font-family: 'Avenir Next LT Pro';
   font-weight: var(--font-demi-bold);
@@ -57,27 +47,6 @@
   width: 14px;
 }
 
-/* Modal styles */
-.modalHeader {
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  width: 100%;
-  justify-content: space-around;
-}
-
-.modalTitle {
-  font-weight: var(--font-heavy);
-  font-size: var(--font-xl);
-  line-height: 25px;
-
-  text-align: center;
-  text-transform: uppercase;
-
-  color: var(--neutral-light-2);
-}
-
-/* TODO: figure out how to do this as .modal .title */
 .modalHeading {
   margin-bottom: var(--unit-2);
 }
@@ -94,6 +63,7 @@
   border-radius: 4px;
   width: 100%;
   padding: var(--unit-3) var(--unit-4);
+  margin-bottom: 128px;
 }
 
 .datePicker :global(.DateInput_input) {
